% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{show_palette}
\alias{show_palette}
\title{Display a named color palette}
\usage{
show_palette(name = "primary")
}
\arguments{
\item{name}{The name of the palette (e.g., "primary", "blacks").}
}
\description{
Display a named color palette
}
