% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinpal-package.R
\docType{package}
\name{vinpal-package}
\alias{vinpal}
\alias{vinpal-package}
\title{Vincent Default Color Palette}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
